HS-520: Separate service into multiple modules
