+++ events/main/src/main/resources/application.yml
@@ -1,0 +1,23 @@
+spring:
+  application:
+    name: horizon-stream-events
+
+  datasource:
+    driver-class-name: org.postgresql.Driver
+    url: jdbc:postgresql://postgres:5432/events
+    username: events
+    password: OpenNMS
+
+  liquibase:
+    change-log: classpath:db/changelog/changelog.xml
+
+  kafka:
+    consumer:
+      group-id: "traps-consumer"
+      auto-offset-reset: earliest
+      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
+      value-deserializer: org.apache.kafka.common.serialization.ByteArrayDeserializer
+    bootstrap-servers: localhost:59092
+
+kafka:
+  topics: "traps"
