+++ events/persistence/src/main/java/org/opennms/horizon/events/persistence/model/EventParameters.java
@@ -1,0 +1,49 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2022 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2022 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.horizon.events.persistence.model;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class EventParameters {
+
+    private List<EventParameter> parameters = new ArrayList<>();
+
+    public List<EventParameter> getParameters() {
+        return parameters;
+    }
+
+    public void setParameters(List<EventParameter> parameters) {
+        this.parameters = parameters;
+    }
+
+    public void addEventParameters(EventParameter eventParameter) {
+        this.parameters.add(eventParameter);
+    }
+}
